Fix MSVC C4267 warning in generated debug-mode WriteFile prologue
len (size_t) was passed to WriteFile's DWORD length param without a
cast in both debug-write prologues, warning on every MSVC debug build.
Diagnostic strings are short so plain truncating cast is safe.

+++ emitter_c.py
@@ -196,7 +196,7 @@ def _prologue_debug_list() -> str:
 	void* h = GetStdHandle( (uint32_t)-12 ); // STD_ERROR_HANDLE
 	if ( h != (void*)(intptr_t)-1 && h != 0 ) {
 		uint32_t written = 0;
-		WriteFile( h, (const uint8_t*)s, len, &written, 0 );
+		WriteFile( h, (const uint8_t*)s, (uint32_t)len, &written, 0 );
 	}
 '''
 		poison_exit = '\tExitProcess( 3 );\n'
@@ -446,7 +446,7 @@ def _prologue_debug_ops() -> str:
 	void* h = GetStdHandle( (uint32_t)-11 ); // STD_OUTPUT_HANDLE
 	if ( h == (void*)(intptr_t)-1 || h == 0 ) return;
 	uint32_t written = 0;
-	WriteFile( h, (const uint8_t*)s, len, &written, 0 );
+	WriteFile( h, (const uint8_t*)s, (uint32_t)len, &written, 0 );
 '''
 		# side-table tracking node allocation - hand-declared HeapAlloc/
 		# HeapFree/GetProcessHeap prototypes (same "repeated, compatible
